refactor response locations (#756)
