+++ r2r/main/api/routes/management.py
@@ -1,19 +1,51 @@
-from fastapi import Depends
+import uuid
+from typing import Optional, Union
 
-from r2r.base import R2RException
+from fastapi import Depends
+from pydantic import BaseModel
+
+from r2r.base import AnalysisTypes, FilterCriteria, R2RException
 
 from ...engine import R2REngine
-from ..requests import (
-    R2RAnalyticsRequest,
-    R2RDeleteRequest,
-    R2RDocumentChunksRequest,
-    R2RDocumentsOverviewRequest,
-    R2RLogsRequest,
-    R2RPrintRelationshipsRequest,
-    R2RUpdatePromptRequest,
-    R2RUsersOverviewRequest,
-)
 from .base_router import BaseRouter
+
+
+class R2RUpdatePromptRequest(BaseModel):
+    name: str
+    template: Optional[str] = None
+    input_types: Optional[dict[str, str]] = {}
+
+
+class R2RDeleteRequest(BaseModel):
+    keys: list[str]
+    values: list[Union[bool, int, str]]
+
+
+class R2RAnalyticsRequest(BaseModel):
+    filter_criteria: FilterCriteria
+    analysis_types: AnalysisTypes
+
+
+class R2RUsersOverviewRequest(BaseModel):
+    user_ids: Optional[list[uuid.UUID]]
+
+
+class R2RDocumentsOverviewRequest(BaseModel):
+    document_ids: Optional[list[uuid.UUID]]
+    user_ids: Optional[list[uuid.UUID]]
+
+
+class R2RDocumentChunksRequest(BaseModel):
+    document_id: uuid.UUID
+
+
+class R2RLogsRequest(BaseModel):
+    log_type_filter: Optional[str] = (None,)
+    max_runs_requested: int = 100
+
+
+class R2RPrintRelationshipsRequest(BaseModel):
+    limit: int = 100
 
 
 class ManagementRouter(BaseRouter):
@@ -185,3 +217,8 @@ class ManagementRouter(BaseRouter):
                     403,
                 )
             return await self.engine.aapp_settings()
+
+
+class R2RExtractionRequest(BaseModel):
+    entity_types: list[str]
+    relations: list[str]
